api/serializers.py
from rest_framework import serializers

from api.models import Photo


class PhotoSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Photo
        fields = ('url', 'id', 'image')