requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.9.4
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
Pillow==6.1.0
protobuf==3.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
